created separate files for each class and fixed host in app.py
