+++ .pre-commit-config.yaml
@@ -3,9 +3,9 @@ repos:
     rev: v2.3.0
     hooks:
     -   id: check-yaml
-    -   id: end-of-file-fixer
+#    -   id: end-of-file-fixer
     -   id: trailing-whitespace
--   repo: https://github.com/psf/black
-    rev: 22.3.0
-    hooks:
-    -   id: black
+#-   repo: https://github.com/psf/black
+#    rev: 22.3.0
+#    #hooks:
+##    -   id: black
