+++ api/app.py
@@ -1,5 +1,6 @@
 from flask import Flask, request, jsonify
-from src.calculate import Np_Math, Basic_Math
+from src.Basic_Math import BasicMath
+from src.Np_Math import NpMath
 from flask import g
 
 app = Flask(__name__)
@@ -25,7 +26,7 @@ def initialize():
     global np_data
     np_data = data["np_data"]
     global math_obj
-    math_obj = Np_Math(np_data)
+    math_obj = NpMath(np_data)
     return jsonify({"result": "Initialized Successfully"})
 
 
@@ -61,4 +62,4 @@ if __name__ == "__main__":
     # run() method of Flask class runs the application
     # on the local development server.
 
-    app.run()
+    app.run(host="0.0.0.0")
